+++ api/dns/record/views.py
@@ -11,6 +11,7 @@ __all__ = ('dns_record_list', 'dns_record')
 def dns_record_list(request, name, data=None):
     """
     List (:http:get:`GET </dns/domain/(name)/record>`) all DNS records which belong to a DNS domain.
+    Delete (:http:delete:`DELETE </dns/domain/(name)/record>`) specified DNS records that belong to a DNS domain.
 
     .. http:get:: /dns/domain/(name)/record
 
@@ -21,6 +22,8 @@ def dns_record_list(request, name, data=None):
             * |DnsAdmin| or |DomainOwner|
         :Asynchronous?:
             * |async-no|
+        :arg name: **required** - Domain name
+        :type name: string
         :arg data.full: Return list of objects with all record details (default: false)
         :type data.full: boolean
         :arg data.order_by: :ref:`Available fields for sorting <order_by>`: ``id``, ``name``, ``type``, ``ttl``, \
@@ -28,6 +31,25 @@ def dns_record_list(request, name, data=None):
         :type data.order_by: string
         :status 200: SUCCESS
         :status 403: Forbidden
+
+    .. http:delete:: /dns/domain/(name)/record
+
+        :DC-bound?:
+            * |dc-yes| - ``domain.dc_bound=true``
+            * |dc-no| - ``domain.dc_bound=false``
+        :Permissions:
+            * |DnsAdmin| or |DomainOwner|
+        :Asynchronous?:
+            * |async-no|
+        :arg name: **required** - Domain name
+        :type name: string
+        :arg data.records: **required** List of DNS record ids to be deleted
+        :type data.records: array
+        :status 200: SUCCESS
+        :status 400: FAILURE
+        :status 403: Forbidden
+        :status 404: Domain not found / Record not found
+        :status 412: Invalid records
     """
     if request.method == 'POST':
         return dns_record(request, name, 0, data=data)