+++ doc/changelog.rst
@@ -17,6 +17,7 @@ Features
 Bugs
 ----
 
+- Documented and implemented hidden DELETE methods for snapshot, backup, DNS records, and IP list API calls - `#237 <https://github.com/erigones/esdc-ce/issues/237>`__
 - Fixed allowed_ips type on all occurrences to list instead of set to enable JSON serialization - `#242 <https://github.com/erigones/esdc-ce/issues/242>`__
 - Fixed bug when deploying VM with dhcp_passthrough network - `#249 <https://github.com/erigones/esdc-ce/issues/249>`__
 - Updated all internal service VM images to be available from the image server and mgmt system - `#244 <https://github.com/erigones/esdc-ce/issues/244>`__